% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dist.R
\name{sample_dist}
\alias{sample_dist}
\title{Sample from a Probability Distribution}
\usage{
sample_dist(n = NULL, dist = NULL, ...)
}
\arguments{
\item{n}{Numeric (or integer), the number of samples to take.}

\item{dist}{A function which samples from a probability distribution. The first
arguement must be the number of samples, additional arguements can be specified with
\code{...}. Possible options are \code{rnorm} (normal), \code{rpois} (Poisson), and
\code{rbinom} (Binomial).}

\item{...}{Additional arguements to pass to the specified \code{dist} function.}
}
\value{
A vector of samples from the specified distribution.
}
\description{
Function brief: Write a function that will draw random numbers
from a given distribution. The function should take one argument for the
number of samples and a second argument which specifies the distribution
(Normal, Poisson or Binomial). The function should be able to handle
additional parameters depending on the distribution chosen, e.g., a \code{lambda}
for Poisson samples.
}
\examples{

## Sample a single time from the normal distribution
sample_dist(1, dist = rnorm, mean = 0, sd = 1)

## Sample multiple times from the poisson distribution
sample_dist(100, dist = rpois, lambda = 2)

}
