% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sample_dist.R
\name{summary.sample_dist}
\alias{summary.sample_dist}
\alias{summary}
\title{Summarising a sample of a probability distribution.}
\usage{
\method{summary}{sample_dist}(object = NULL, plot = TRUE, ...)
}
\arguments{
\item{object}{A numeric vector of class "sample_dist".}

\item{plot}{Logical, defaults to \code{TRUE}. Produces a summary plot.}

\item{...}{Pass additional arguements to summary function.}
}
\value{
Prints a summary table to the terminal and a summary plot.
}
\description{
\code{summary} method for class "sample_dist".
}
\examples{
## Sample from a normal distribution
x <- sample_dist(100, dist = rnorm)

## Create a summary
summary(x)
}
\seealso{
summary
}
