% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sample_dist.R
\name{summary.sample_dist}
\alias{summary.sample_dist}
\title{Summarising a sample of a probability distribution.}
\usage{
## S3 method for class
summary(sample_vect, plot = TRUE)
}
\arguments{
\item{sample_vect}{A numeric vector of class "sample_dist".}

\item{plot}{Logical, defaults to \code{TRUE}. Produces a summary plot.}
}
\value{
Prints a summary table to the terminal
}
\description{
\code{summary} method for class "sample_dist".
}
\examples{
## Sample from a normal distribution
x <- sample_dist(100, dist = rnorm)

## Create a summary
summary(x)
}
\seealso{
summary
}
